add ZERO_CROSSING as OST mode to track fricatives
Added ZERO_CROSSING to track fricatives based on zero-crossing rate in the speech signal. Two arguments are provided. The first is a thershold for the zero crossing rate, crossing the threshold will increase the ost_stat by 1. The OST stat will remain at that level as long as the zero crossing rate stays above the threshold. If it returns below the threshold; the stat is increased by 1 only if the total duration was longer than the OST mode's second argument.